Move config-session files compatibility tests to new test class
