+++ tests/CMakeLists.txt
@@ -56,4 +56,5 @@ endfunction()
 
 SETUP_TESTS(
     test_jsonloggercontentloader
+    test_jsonconfigfiles
     )
