+++ tests/test_jsonloggercontentloader.cpp
@@ -17,33 +17,6 @@ void test_jsonloggercontentloader::noSessionSetEmptyAvailableContentSets()
     QVERIFY(loader.getAvailableContentSets().isEmpty());
 }
 
-
-void test_jsonloggercontentloader::compareSessionLogConfigFileCountEqual()
-{
-    QStringList sessionFileList = QString(SESSION_FILES).split(",");
-    QStringList loggerFileList = QString(LOGGER_CONFIG_FILES).split(",");
-    QCOMPARE(sessionFileList.count(), loggerFileList.count());
-}
-
-static void toBaseName(QStringList &fileList)
-{
-    for(auto& file : fileList) {
-        QFileInfo fi(file);
-        file = fi.baseName();
-    }
-}
-
-void test_jsonloggercontentloader::compareSessionLogConfigFileBasenamesEqual()
-{
-    QStringList sessionFileList = QString(SESSION_FILES).split(",");
-    sessionFileList.sort();
-    toBaseName(sessionFileList);
-    QStringList loggerFileList = QString(LOGGER_CONFIG_FILES).split(",");
-    loggerFileList.sort();
-    toBaseName(loggerFileList);
-    QCOMPARE(sessionFileList, loggerFileList);
-}
-
 void test_jsonloggercontentloader::testSimpleAvailContentSets()
 {
     JsonLoggerContentLoader loader;
